requirements.txt
# Enhanced Window Characteristic Document Agent Requirements

# Core PDF and Image Processing
pdf2image>=3.1.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# Advanced Document Processing
docling>=1.0.0
docling-core>=1.0.0

# Azure OpenAI Integration
langchain-openai>=0.1.0
langchain-core>=0.2.0

# Environment Management
python-dotenv>=1.0.0

# Web Interface
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0

# Utilities
requests>=2.31.0

# Optional: For better performance
# watchdog>=3.0.0